CI: Avoid error with `cat` if there are multiple output log files per test
If a test is run with multiple (different) numbers of parallel processes,
a separate set of `test-stderr*.log` and `test-stdout*.log` files is
created for each number of parallel processes. If one of these tests
fails for some reason, `cat` is essentially called with multiple file
arguments (which it doesn't support).

Avoid that error by iterating through the list of `test-stderr*.log` and
`test-stdout*.log` files.