Bubble trouble (#635)
Fix the handling of bubble DOFs

Fix the creation of permutation vector for bubble DOFs. Now one should get
indices for bubble DOFs only when the solver has
"Bubbles In Global System = True" (the default) together with

(i) either an "Element" command leading to bubble DOFs
(ii) or the command "Bubbles = True".

In the case of "Bubbles = True" the count of bubble DOFs cannot be controlled in
a solverwise manner (the maximal count of available bubble indices is returned).
Otherwise the count of bubble DOFs is determined in a solverwise manner, so
for the best control of bubble DOFs it is recommended to use an Element command.

Change the subroutines GetElementNOFDOFs, GetElementNOFBDOFs and
GetElementDOFs so that they are consistent with the handling of bubbles
explained in the first item.

A test case NaturalConvectionRestartCycle has suffered from having permutations
of wrong sizes and also needed a fix of norm.

In addition to these changes, this commit contains an unrelated fix for
BlockSolve to avoid processing outside array bounds.