CI: Disable building with OpenMP on Apple Silicon
Building with OpenMP causes some test failures for macOS on Apple
Silicon, see:
https://github.com/ElmerCSC/elmerfem/pull/634#issuecomment-2613840076
The affected tests currently are: `H1BasisEvaluation` and
`SD_H1BasisEvaluation`.

For the time being, disable building with OpenMP on Apple Silicon in CI.

OpenMP is needed as a transitional build dependency of the SuiteSparse
package from Homebrew. So, also disable building with CHOLMOD.